feat(shelter-operator): put the real date controls in the reports toolbar
ReportFilterBar's date field and preset dropdown were placeholders while its
export flow became real. Swap in DateRangeCalendar and DateRangePresetDropdown
and keep the export button and modal untouched. DateRangeFilterBar duplicated
this toolbar with a dead export button, so it goes.

Its test mocked jotai wholesale, which the real controls can't work against;
seed the store instead.

Also define --font-sans in the shared theme. Tailwind's namespace is --font-*,
so --font-family-primary never overrode it and every `font-sans` element fell
back to the system stack, in the app and in Storybook.
