+++ libs/react/shelter-operator/src/lib/components/date-range-filter/index.ts
@@ -15,5 +15,3 @@ export { DateRangePresetDropdown } from './DateRangePresetDropdown';
 export type { DateRangePresetDropdownProps } from './DateRangePresetDropdown';
 export { DateRangeCalendar } from './DateRangeCalendar';
 export type { DateRangeCalendarProps } from './DateRangeCalendar';
-export { DateRangeFilterBar } from './DateRangeFilterBar';
-export type { DateRangeFilterBarProps } from './DateRangeFilterBar';
