+++ libs/tailwind/src/css/theme.css
@@ -1,5 +1,9 @@
 @theme {
   /* Font family */
+  /* Tailwind's namespace is --font-*, so --font-family-primary does not override
+     the built-in --font-sans. Set it explicitly, otherwise `font-sans` falls back
+     to the system stack instead of the brand font. */
+  --font-sans: 'Poppins', ui-sans-serif, system-ui, sans-serif;
   --font-family-primary: 'Poppins', ui-sans-serif, system-ui, sans-serif;
 
   /* Color palette */
